Update site's meta description via main settings
The description field in main settings does not update the site's meta description tag. It only updates the "site_description" configuration value which is never used to generate site's meta description. Instead, meta description is generated with "page_default_description" language value of navigation. This fix updates the language value with the description input every time it is changed. Also, it shows the language value as input.