Change RegExp
Allow any kind of letter from any language, any kind of whitespace or invisible separator and any kind of numeric character to support international usernames. Whitespaces are added to support usernames that consists in two parts, such as "John Doe" in order to be able to enter first and last name. In this way, username and realname fields are not needed anymore.